feat: 친구 추천에 사용되는 DTO 분리 (#324)
